+++ wpilibc/athena/include/PowerDistributionPanel.h
@@ -21,12 +21,12 @@ class PowerDistributionPanel : public SensorBase, public LiveWindowSendable {
   PowerDistributionPanel();
   explicit PowerDistributionPanel(uint8_t module);
 
-  double GetVoltage() const;
-  double GetTemperature() const;
-  double GetCurrent(uint8_t channel) const;
-  double GetTotalCurrent() const;
-  double GetTotalPower() const;
-  double GetTotalEnergy() const;
+  float GetVoltage() const;
+  float GetTemperature() const;
+  float GetCurrent(uint8_t channel) const;
+  float GetTotalCurrent() const;
+  float GetTotalPower() const;
+  float GetTotalEnergy() const;
   void ResetTotalEnergy();
   void ClearStickyFaults();
