+++ wpilibc/shared/include/Utility.h
@@ -45,8 +45,8 @@ bool wpi_assertNotEqual_impl(int valueA, int valueB, const char* valueAString,
 
 void wpi_suspendOnAssertEnabled(bool enabled);
 
-uint16_t GetFPGAVersion();
-uint32_t GetFPGARevision();
+int32_t GetFPGAVersion();
+int64_t GetFPGARevision();
 uint64_t GetFPGATime();
 bool GetUserButton();
 std::string GetStackTrace(uint32_t offset);
